# src/routes/swagger.json
{
  "openapi": "3.0.0",
  "components": {
...
    "schemas": {
...
      "Record_string.unknown_": {
        "properties": {},
        "additionalProperties": {},
        "type": "object",
        "description": "Construct a type with a set of properties K of type T"
      },
...
      "RecordId": {
        "type": "string",
        "example": "821f9b26-ad04-4f56-89b6-e2ef9c72b36e"
      },
...
      "JsonValue": {
        "anyOf": [
          {
            "type": "string"
          },
          {
            "type": "number",
            "format": "double"
          },
          {
            "type": "boolean"
          },
          {
            "$ref": "#/components/schemas/JsonObject"
          },
          {
            "$ref": "#/components/schemas/JsonArray"
          }
        ],
        "nullable": true
      },
      "JsonObject": {
        "properties": {},
        "type": "object",
        "additionalProperties": {
          "$ref": "#/components/schemas/JsonValue"
        }
      },
      "JsonArray": {
        "items": {
          "$ref": "#/components/schemas/JsonValue"
        },
        "type": "array"
      },
      "Pick_JwsGeneralFormat.Exclude_keyofJwsGeneralFormat.payload__": {
        "properties": {
          "header": {
            "$ref": "#/components/schemas/Record_string.unknown_",
            "description": "unprotected header"
          },
          "signature": {
            "type": "string",
            "description": "Base64url encoded signature"
          },
          "protected": {
            "type": "string",
            "description": "Base64url encoded protected header"
          }
        },
        "required": [
          "header",
          "signature",
          "protected"
        ],
        "type": "object",
        "description": "From T, pick a set of properties whose keys are in the union K"
      },
      "Omit_JwsGeneralFormat.payload_": {
        "$ref": "#/components/schemas/Pick_JwsGeneralFormat.Exclude_keyofJwsGeneralFormat.payload__",
        "description": "Construct a type with the properties of T except for those in type K."
      },
      "JwsDetachedFormat": {
        "$ref": "#/components/schemas/Omit_JwsGeneralFormat.payload_"
      },
      "JwsFlattenedDetachedFormat": {
        "properties": {
          "signatures": {
            "items": {
              "$ref": "#/components/schemas/JwsDetachedFormat"
            },
            "type": "array"
          }
        },
        "required": [
          "signatures"
        ],
        "type": "object",
        "additionalProperties": false
      },
      "AttachmentData": {
        "description": "A JSON object that gives access to the actual content of the attachment",
        "properties": {
          "base64": {
            "type": "string",
            "description": "Base64-encoded data, when representing arbitrary content inline instead of via links. Optional."
          },
          "json": {
            "$ref": "#/components/schemas/JsonValue",
            "description": "Directly embedded JSON data, when representing content inline instead of via links, and when the content is natively conveyable as JSON. Optional."
          },
          "links": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "description": "A list of zero or more locations at which the content may be fetched. Optional."
          },
          "jws": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JwsDetachedFormat"
              },
              {
                "$ref": "#/components/schemas/JwsFlattenedDetachedFormat"
              }
            ],
            "description": "A JSON Web Signature over the content of the attachment. Optional."
          },
          "sha256": {
            "type": "string",
            "description": "The hash of the content. Optional."
          }
        },
        "type": "object",
        "additionalProperties": false
      },
      "Attachment": {
        "description": "Represents DIDComm attachment\nhttps://github.com/hyperledger/aries-rfcs/blob/master/concepts/0017-attachments/README.md",
        "properties": {
          "id": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "description": "An optional human-readable description of the content."
          },
          "filename": {
            "type": "string",
            "description": "A hint about the name that might be used if this attachment is persisted as a file. It is not required, and need not be unique. If this field is present and mime-type is not, the extension on the filename may be used to infer a MIME type."
          },
          "mimeType": {
            "type": "string",
            "description": "Describes the MIME type of the attached content. Optional but recommended."
          },
          "lastmodTime": {
            "type": "string",
            "format": "date-time",
            "description": "A hint about when the content in this attachment was last modified."
          },
          "byteCount": {
            "type": "number",
            "format": "double",
            "description": "Optional, and mostly relevant when content is included by reference instead of by value. Lets the receiver guess how expensive it will be, in time, bandwidth, and storage, to fully fetch the attachment."
          },
          "data": {
            "$ref": "#/components/schemas/AttachmentData"
          }
        },
        "required": [
          "id",
          "data"
        ],
        "type": "object",
        "additionalProperties": false
      },
...
      "ProtocolVersion": {
        "enum": [
          "v1",
          "v2"
        ],
        "type": "string"
      },
      "CredentialPreviewAttributeOptions": {
        "properties": {
          "name": {
            "type": "string"
          },
          "mimeType": {
            "type": "string"
          },
          "value": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "value"
        ],
        "type": "object",
        "additionalProperties": false
      },
      "LinkedAttachment": {
        "properties": {
          "attributeName": {
            "type": "string",
            "description": "The name that will be used to generate the linked credential"
          },
          "attachment": {
            "$ref": "#/components/schemas/Attachment",
            "description": "The attachment that needs to be linked to the credential"
          }
        },
        "required": [
          "attributeName",
          "attachment"
        ],
        "type": "object",
        "additionalProperties": false
      },
...
      "W3cIssuerOptions": {
        "description": "TODO: check how to support arbitrary data in class",
        "properties": {
          "id": {
            "type": "string"
          }
        },
        "required": [
          "id"
        ],
        "type": "object",
        "additionalProperties": false
      },
      "SingleOrArray_JsonObject_": {
        "anyOf": [
          {
            "$ref": "#/components/schemas/JsonObject"
          },
          {
            "items": {
              "$ref": "#/components/schemas/JsonObject"
            },
            "type": "array"
          }
        ]
      },
      "JsonCredential": {
        "properties": {
          "@context": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "$ref": "#/components/schemas/JsonObject"
              }
            ]
          },
          "id": {
            "type": "string"
          },
          "type": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "issuer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "$ref": "#/components/schemas/W3cIssuerOptions"
              }
            ]
          },
          "issuanceDate": {
            "type": "string"
          },
          "expirationDate": {
            "type": "string"
          },
          "credentialSubject": {
            "$ref": "#/components/schemas/SingleOrArray_JsonObject_"
          },
          "prettyVc": {}
        },
        "required": [
          "@context",
          "type",
          "issuer",
          "issuanceDate",
          "credentialSubject"
        ],
        "type": "object",
        "additionalProperties": {}
      },
      "JsonLdCredentialDetailFormat": {
        "description": "Format for creating a jsonld proposal, offer or request.",
        "properties": {
          "credential": {
            "$ref": "#/components/schemas/JsonCredential"
          },
          "options": {
            "properties": {
              "proofType": {
                "type": "string"
              },
              "proofPurpose": {
                "type": "string"
              }
            },
            "required": [
              "proofType",
              "proofPurpose"
            ],
            "type": "object"
          }
        },
        "required": [
          "credential",
          "options"
        ],
        "type": "object",
        "additionalProperties": false
      },
...
      "AutoAcceptCredential": {
        "description": "Typing of the state for auto acceptance",
        "enum": [
          "always",
          "contentApproved",
          "never"
        ],
        "type": "string"
      },
...
      "AnonCredsOfferCredentialFormat": {
        "description": "This defines the module payload for calling CredentialsApi.offerCredential\nor CredentialsApi.negotiateProposal",
        "properties": {
          "credentialDefinitionId": {
            "type": "string"
          },
          "revocationRegistryDefinitionId": {
            "type": "string"
          },
          "revocationRegistryIndex": {
            "type": "number",
            "format": "double"
          },
          "attributes": {
            "items": {
              "$ref": "#/components/schemas/CredentialPreviewAttributeOptions"
            },
            "type": "array"
          },
          "linkedAttachments": {
            "items": {
              "$ref": "#/components/schemas/LinkedAttachment"
            },
            "type": "array"
          }
        },
        "required": [
          "credentialDefinitionId",
          "attributes"
        ],
        "type": "object",
        "additionalProperties": false
      },
      "CredentialFormatPayload_CredentialFormats.createOffer_": {
        "properties": {
          "indy": {
            "$ref": "#/components/schemas/AnonCredsOfferCredentialFormat"
          },
          "jsonld": {
            "$ref": "#/components/schemas/JsonLdCredentialDetailFormat"
          },
          "anoncreds": {
            "$ref": "#/components/schemas/AnonCredsOfferCredentialFormat"
          }
        },
        "type": "object",
        "description": "Get the payload for a specific method from a list of CredentialFormat interfaces and a method"
      },
      "CreateOfferOptions": {
        "properties": {
          "protocolVersion": {
            "$ref": "#/components/schemas/ProtocolVersion"
          },
          "connectionId": {
            "$ref": "#/components/schemas/RecordId"
          },
          "credentialFormats": {
            "$ref": "#/components/schemas/CredentialFormatPayload_CredentialFormats.createOffer_"
          },
          "autoAcceptCredential": {
            "$ref": "#/components/schemas/AutoAcceptCredential"
          },
          "comment": {
            "type": "string"
          },
          "goalCode": {
            "type": "string"
          },
          "goal": {
            "type": "string"
          }
        },
        "required": [
          "protocolVersion",
          "connectionId",
          "credentialFormats"
        ],
        "type": "object",
        "additionalProperties": false
      },
...
      "CreateCredentialOfferResponse": {
        "description": "Create offer response",
        "properties": {
          "offerId": {
            "type": "string"
          },
          "credential_offer_url": {
            "type": "string"
          },
          "preAuthorizedCode": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string"
          }
        },
        "required": [
          "offerId",
          "credential_offer_url",
          "preAuthorizedCode",
          "expiresAt"
        ],
        "type": "object",
        "additionalProperties": false
      },
      "OfferCredentialTemplate": {
        "description": "Credential template item for offers",
        "properties": {
          "type": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "schemaId": {
            "type": "string"
          },
          "claimsTemplate": {},
          "format": {
            "type": "string",
            "enum": [
              "jwt_vc",
              "sd_jwt"
            ]
          },
          "credentialDefinitionId": {
            "type": "string"
          },
          "issuerDid": {
            "type": "string"
          }
        },
        "required": [
          "type"
        ],
        "type": "object",
        "additionalProperties": false
      },
      "CreateCredentialOfferRequest": {
        "description": "Create credential offer request",
        "properties": {
          "credentials": {
            "items": {
              "$ref": "#/components/schemas/OfferCredentialTemplate"
            },
            "type": "array"
          },
          "issuerDid": {
            "type": "string"
          },
          "expiresIn": {
            "type": "number",
            "format": "double"
          }
        },
        "required": [
          "credentials"
        ],
        "type": "object",
        "additionalProperties": false
      },
      "TokenResponseBody": {
        "description": "Token response with issued credential",
        "properties": {
          "verifiableCredential": {
            "type": "string"
          },
          "credentialId": {
            "type": "string"
          }
        },
        "required": [
          "verifiableCredential",
          "credentialId"
        ],
        "type": "object",
        "additionalProperties": false
      },
      "TokenRequestBody": {
        "description": "Token exchange request (pre-authorized code)",
        "properties": {
          "grant_type": {
            "type": "string"
          },
          "pre_authorized_code": {
            "type": "string"
          },
          "subject_did": {
            "type": "string"
          },
          "format": {
            "type": "string",
            "enum": [
              "jwt_vc",
              "sd_jwt"
            ]
          }
        },
        "required": [
          "grant_type",
          "pre_authorized_code",
          "subject_did"
        ],
        "type": "object",
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header"
      },
      "jwt": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "info": {
    "title": "credo-controller",
    "version": "2.0.0",
    "description": "Rest endpoint wrapper for using your agent over HTTP",
    "license": {
      "name": "Apache-2.0"
    },
    "contact": {}
  },
  "paths": {
...
    "/didcomm/credentials/create-offer": {
      "post": {
        "operationId": "CreateOffer",
        "responses": {
          "200": {
            "description": "AgentMessage, CredentialExchangeRecord",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "type": "object"
                },
                "examples": {
                  "Example 1": {
                    "value": {
                      "_tags": {
                        "state": "offer-sent",
                        "threadId": "82701488-b43c-4d7b-9244-4bb204a7ae26",
                        "connectionId": "ac6d0fdd-0db8-4f52-8a3d-de7ff8ddc14b"
                      },
                      "metadata": {
                        "_internal/w3cCredential": {
                          "credentialDefinitionId": "q7ATwTYbQDgiigVijUAej:3:CL:318187:latest",
                          "schemaId": "q7ATwTYbQDgiigVijUAej:2:Employee Badge:1.0"
                        }
                      },
                      "credentials": [],
                      "id": "821f9b26-ad04-4f56-89b6-e2ef9c72b36e",
                      "createdAt": "2022-01-01T00:00:00.000Z",
                      "connectionId": "ac6d0fdd-0db8-4f52-8a3d-de7ff8ddc14b",
                      "threadId": "82701488-b43c-4d7b-9244-4bb204a7ae26",
                      "credentialAttributes": [],
                      "protocolVersion": "v1"
                    }
                  }
                }
              }
            },
            "headers": {
              "x-correlation-id": {
                "description": "Correlation ID for tracing across systems",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "description": "Initiate a new credential exchange as issuer by creating a credential offer\nwithout specifying a connection id",
        "tags": [
          "DIDComm - Credentials"
        ],
        "security": [
          {
            "jwt": [
              "tenant",
              "dedicated"
            ]
          }
        ],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOfferOptions"
              }
            }
          }
        }
      }
    },
...
    "/oidc/issuer/credential-offers": {
      "post": {
        "operationId": "CreateCredentialOffer",
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateCredentialOfferResponse"
                }
              }
            },
            "headers": {
              "x-correlation-id": {
                "description": "Correlation ID for tracing across systems",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "description": "Create a credential offer (pre-authorized code flow) - placeholder implementation.",
        "tags": [
          "OIDC4VC"
        ],
        "security": [
          {
            "jwt": [
              "tenant"
            ]
          }
        ],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCredentialOfferRequest"
              }
            }
          }
        }
      }
    },
    "/oidc/token": {
      "post": {
        "operationId": "Token",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponseBody"
                }
              }
            },
            "headers": {
              "x-correlation-id": {
                "description": "Correlation ID for tracing across systems",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "description": "Token exchange for pre-authorized code -> issues VC (JWT or SD-JWT) - placeholder.",
        "tags": [
          "OIDC4VC"
        ],
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenRequestBody"
              }
            }
          }
        }
      }
    },
    "/oidc/issuer/credentials/{id}": {
      "get": {
        "operationId": "GetCredential",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "headers": {
              "x-correlation-id": {
                "description": "Correlation ID for tracing across systems",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "description": "Get issued credential metadata",
        "tags": [
          "OIDC4VC"
        ],
        "security": [
          {
            "jwt": [
              "tenant"
            ]
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/oidc/issuer/credentials/{id}/revoke": {
      "post": {
        "operationId": "RevokeCredential",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "headers": {
              "x-correlation-id": {
                "description": "Correlation ID for tracing across systems",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "description": "Revoke an issued credential (simple flag)",
        "tags": [
          "OIDC4VC"
        ],
        "security": [
          {
            "jwt": [
              "tenant"
            ]
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/oidc/issuer/credentials": {
      "get": {
        "operationId": "ListCredentials",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "items": {},
                  "type": "array"
                }
              }
            },
            "headers": {
              "x-correlation-id": {
                "description": "Correlation ID for tracing across systems",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "description": "List issued credentials with optional filtering",
        "tags": [
          "OIDC4VC"
        ],
        "security": [
          {
            "jwt": [
              "tenant"
            ]
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "subject",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "issuer",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    }
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}